Move type ErrorsAndWarnings to InputGridRow
